+++ vscode_task_runner2/models/execution_level.py
@@ -1,8 +1,8 @@
 from pydantic import BaseModel
 
-from vscode_task_runner2.models.task import DependsOrderEnum, Task
+from vscode_task_runner2.models.task import DependsOrder, Task
 
 
 class ExecutionLevel(BaseModel):
-    order: DependsOrderEnum
+    order: DependsOrder
     tasks: list[Task]
